serving_benchmarks/requirements.txt
PyYAML
# Optional (uncomment and pin in your environment as needed):
# torch
# numpy
# matplotlib